fix: UI improvments for self deleting messages in group
