+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/SystemMessageItem.kt
@@ -67,6 +67,7 @@ import com.wire.android.ui.home.conversations.model.UIMessageContent.SystemMessa
 import com.wire.android.ui.theme.wireColorScheme
 import com.wire.android.ui.theme.wireTypography
 import com.wire.android.util.ui.UIText
+import com.wire.android.util.ui.annotatedText
 import com.wire.android.util.ui.stringWithStyledArgs
 import com.wire.android.util.ui.toUIText
 
@@ -334,16 +335,17 @@ fun SystemMessage.annotatedString(
         is SystemMessage.HistoryLost -> arrayOf()
         is SystemMessage.ConversationMessageTimerActivated -> arrayOf(
             author.asString(res),
-            res.getString(R.string.label_system_message_activated),
             selfDeletionDuration.longLabel.asString(res)
         )
-
-        is SystemMessage.ConversationMessageTimerDeactivated -> arrayOf(
-            author.asString(res),
-            res.getString(R.string.label_system_message_deactivated)
-        )
+        is SystemMessage.ConversationMessageTimerDeactivated -> arrayOf(author.asString(res))
     }
-    return res.stringWithStyledArgs(stringResId, normalStyle, boldStyle, normalColor, boldColor, *args)
+
+    // TODO
+//    return if(this is SystemMessage.ConversationMessageTimerActivated || this is SystemMessage.ConversationMessageTimerDeactivated) {
+       return res.annotatedText(stringResId, normalStyle, boldStyle, normalColor, boldColor, *args)
+//    } else {
+//        res.stringWithStyledArgs(stringResId, normalStyle, boldStyle, normalColor, boldColor, *args)
+//    }
 }
 
 private const val EXPANDABLE_THRESHOLD = 4
