+++ app/src/test/kotlin/com/wire/android/framework/TestMessage.kt
@@ -105,6 +105,15 @@ object TestMessage {
         isSelfMessage = false
     )
 
+    val SYSTEM_MESSAGE = Message.System(
+        id = "messageID",
+        content = MessageContent.MissedCall,
+        conversationId = ConversationId("convo-id", "convo.domain"),
+        date = "some-date",
+        senderUserId = UserId("user-id", "domain"),
+        status = Message.Status.SENT
+    )
+
     val MEMBER_REMOVED_MESSAGE = Message.System(
         id = "messageID",
         content = MessageContent.MemberChange.Removed(listOf(UserId("user-id", "domain"))),
