+++ app/src/test/kotlin/com/wire/android/mapper/RegularMessageContentMapperTest.kt
@@ -20,24 +20,16 @@
 
 package com.wire.android.mapper
 
-import android.content.res.Resources
 import com.wire.android.config.CoroutineTestExtension
-import com.wire.android.config.TestDispatcherProvider
 import com.wire.android.framework.FakeKaliumFileSystem
 import com.wire.android.framework.TestConversation
 import com.wire.android.framework.TestMessage
 import com.wire.android.framework.TestMessage.buildAssetMessage
-import com.wire.android.framework.TestUser
-import com.wire.android.ui.home.conversations.model.UIMessageContent
 import com.wire.android.ui.home.conversations.model.UIMessageContent.AssetMessage
 import com.wire.android.ui.home.conversations.model.UIMessageContent.PreviewAssetMessage
-import com.wire.android.ui.home.conversations.model.UIMessageContent.SystemMessage
-import com.wire.android.ui.home.conversations.name
 import com.wire.android.util.time.ISOFormatter
 import com.wire.android.util.ui.UIText
 import com.wire.android.util.ui.WireSessionImageLoader
-import com.wire.kalium.logic.data.conversation.Conversation.Member
-import com.wire.kalium.logic.data.conversation.MemberDetails
 import com.wire.kalium.logic.data.id.QualifiedID
 import com.wire.kalium.logic.data.message.AssetContent
 import com.wire.kalium.logic.data.message.AssetContent.AssetMetadata
@@ -46,7 +38,6 @@ import com.wire.kalium.logic.data.message.MessageContent
 import com.wire.kalium.logic.data.user.ConnectionState
 import com.wire.kalium.logic.data.user.OtherUser
 import com.wire.kalium.logic.data.user.UserAvailabilityStatus
-import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.data.user.type.UserType
 import com.wire.kalium.logic.feature.asset.GetMessageAssetUseCase
 import com.wire.kalium.logic.feature.asset.MessageAssetResult
@@ -67,34 +58,7 @@ import org.junit.jupiter.api.extension.ExtendWith
 
 @OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(CoroutineTestExtension::class)
-class EncodedMessageContentMapperTest {
-
-    @Test
-    fun givenMemberDetails_whenMappingToSystemMessageMemberName_thenCorrectValuesShouldBeReturned() = runTest {
-        // Given
-        val (arrangement, mapper) = Arrangement().arrange()
-        val selfMemberDetails = TestUser.MEMBER_SELF
-        val deletedMemberDetails = TestUser.MEMBER_OTHER.copy(TestUser.OTHER_USER.copy(name = null))
-        val otherMemberDetails = TestUser.MEMBER_OTHER
-        // When
-        val selfName = mapper.toSystemMessageMemberName(selfMemberDetails.user, MessageContentMapper.SelfNameType.NameOrDeleted)
-        val selfResLower = mapper.toSystemMessageMemberName(selfMemberDetails.user, MessageContentMapper.SelfNameType.ResourceLowercase)
-        val selfResTitle = mapper.toSystemMessageMemberName(selfMemberDetails.user, MessageContentMapper.SelfNameType.ResourceTitleCase)
-        val deleted = mapper.toSystemMessageMemberName(deletedMemberDetails.user, MessageContentMapper.SelfNameType.NameOrDeleted)
-        val otherName = mapper.toSystemMessageMemberName(otherMemberDetails.user, MessageContentMapper.SelfNameType.NameOrDeleted)
-        // Then
-        assertTrue(selfName is UIText.DynamicString && selfName.value == selfMemberDetails.name)
-        assertTrue(
-            selfResLower is UIText.StringResource
-                    && selfResLower.resId == arrangement.messageResourceProvider.memberNameYouLowercase
-        )
-        assertTrue(
-            selfResTitle is UIText.StringResource
-                    && selfResTitle.resId == arrangement.messageResourceProvider.memberNameYouTitlecase
-        )
-        assertTrue(deleted is UIText.StringResource && deleted.resId == arrangement.messageResourceProvider.memberNameDeleted)
-        assertTrue(otherName is UIText.DynamicString && otherName.value == otherMemberDetails.name)
-    }
+class RegularMessageContentMapperTest {
 
     @Test
     fun givenTextOrNullContent_whenMappingToTextMessageContent_thenCorrectValuesShouldBeReturned() = runTest {
@@ -117,64 +81,6 @@ class EncodedMessageContentMapperTest {
                         (messageBody.message as UIText.StringResource).resId == arrangement.messageResourceProvider.sentAMessageWithContent
             )
         }
-    }
-
-    @Test
-    fun givenServerContent_whenMappingToUIMessageContent_thenCorrectValuesShouldBeReturned() = runTest {
-        // Given
-        val (arrangement, mapper) = Arrangement().arrange()
-        val userId1 = UserId("user-id1", "user-domain")
-        val userId2 = UserId("user-id2", "user-domain")
-        val userId3 = UserId("user-id3", "user-domain")
-        val contentLeft = MessageContent.MemberChange.Removed(listOf(userId1))
-        val contentRemoved = MessageContent.MemberChange.Removed(listOf(userId2))
-        val contentAdded = MessageContent.MemberChange.Added(listOf(userId2, userId3))
-        val contentAddedSelf = MessageContent.MemberChange.Added(listOf(userId1))
-        val member1 = TestUser.MEMBER_OTHER.copy(TestUser.OTHER_USER.copy(id = userId1))
-        val member2 = TestUser.MEMBER_OTHER.copy(TestUser.OTHER_USER.copy(id = userId2))
-        val member3 = TestUser.MEMBER_OTHER.copy(TestUser.OTHER_USER.copy(id = userId3))
-        val missedCallMessage = TestMessage.MISSED_CALL_MESSAGE
-        val selfCaller = MemberDetails(TestUser.SELF_USER.copy(id = missedCallMessage.senderUserId), Member.Role.Admin)
-        val otherCallerInfo = (member1.user as OtherUser).copy(id = missedCallMessage.senderUserId)
-        val otherCaller = member1.copy(user = otherCallerInfo)
-        // When
-        val resultContentLeft = mapper.mapMemberChangeMessage(contentLeft, userId1, listOf(member1.user))
-        val resultContentRemoved = mapper.mapMemberChangeMessage(contentRemoved, userId1, listOf(member1.user, member2.user))
-        val resultContentAdded = mapper.mapMemberChangeMessage(contentAdded, userId1, listOf(member1.user, member2.user, member3.user))
-        val resultContentAddedSelf = mapper.mapMemberChangeMessage(contentAddedSelf, userId1, listOf(member1.user))
-        val resultMyMissedCall = mapper.fromMessage(missedCallMessage, listOf(selfCaller.user))
-        val resultOtherMissedCall = mapper.fromMessage(missedCallMessage, listOf(otherCaller.user))
-        // Then
-        assertTrue(
-            resultContentLeft is SystemMessage.MemberLeft &&
-                    resultContentLeft.author.asString(arrangement.resources) == member1.name
-        )
-        assertTrue(
-            resultContentRemoved is SystemMessage.MemberRemoved &&
-                    resultContentRemoved.author.asString(arrangement.resources) == member1.name &&
-                    resultContentRemoved.memberNames.size == 1 &&
-                    resultContentRemoved.memberNames[0].asString(arrangement.resources) == member2.name
-
-        )
-        assertTrue(
-            resultContentAdded is SystemMessage.MemberAdded &&
-                    resultContentAdded.author.asString(arrangement.resources) == member1.name &&
-                    resultContentAdded.memberNames.size == 2 &&
-                    resultContentAdded.memberNames[0].asString(arrangement.resources) == member2.name &&
-                    resultContentAdded.memberNames[1].asString(arrangement.resources) == member3.name
-        )
-        assertTrue(
-            resultContentAddedSelf is SystemMessage.MemberJoined &&
-                    resultContentAddedSelf.author.asString(arrangement.resources) == member1.name
-        )
-        assertTrue(
-            resultOtherMissedCall is SystemMessage.MissedCall &&
-                    resultOtherMissedCall.author.asString(arrangement.resources) == TestUser.OTHER_USER.name
-        )
-        assertTrue(
-            resultMyMissedCall is SystemMessage.MissedCall &&
-                    (resultMyMissedCall.author as UIText.StringResource).resId == arrangement.messageResourceProvider.memberNameYouTitlecase
-        )
     }
 
     @Test
@@ -286,29 +192,6 @@ class EncodedMessageContentMapperTest {
         }
     }
 
-    @Test
-    fun givenMessagesWithDifferentVisibilities_whenMappingToUIMessageContent_thenCorrectValuesShouldBeReturned() = runTest {
-        // Given
-        val (_, mapper) = Arrangement().arrange()
-        val visibleMessage = TestMessage.TEXT_MESSAGE.copy(visibility = Message.Visibility.VISIBLE)
-        val deletedMessage = TestMessage.TEXT_MESSAGE.copy(
-            visibility = Message.Visibility.DELETED,
-            content = MessageContent.Text("")
-        )
-        val hiddenMessage = TestMessage.TEXT_MESSAGE.copy(
-            visibility = Message.Visibility.HIDDEN,
-            content = MessageContent.Text("")
-        )
-        // When
-        val resultContentVisible = mapper.fromMessage(visibleMessage, listOf())
-        val resultContentDeleted = mapper.fromMessage(deletedMessage, listOf())
-        val resultContentHidden = mapper.fromMessage(hiddenMessage, listOf())
-        // Then
-        assertTrue(resultContentVisible != null)
-        assertTrue(resultContentDeleted == UIMessageContent.Deleted)
-        assertTrue(resultContentHidden == null)
-    }
-
     private class Arrangement {
 
         @MockK
@@ -318,15 +201,10 @@ class EncodedMessageContentMapperTest {
         lateinit var messageResourceProvider: MessageResourceProvider
 
         @MockK
-        lateinit var resources: Resources
-
-        @MockK
         private lateinit var wireSessionImageLoader: WireSessionImageLoader
 
-        val testDispatcher = TestDispatcherProvider()
-
         private val messageContentMapper by lazy {
-            MessageContentMapper(messageResourceProvider, wireSessionImageLoader, ISOFormatter())
+            RegularMessageMapper(messageResourceProvider, wireSessionImageLoader, ISOFormatter())
         }
 
         init {
